+++ govtool/frontend/src/services/requests/metadataValidation/postValidate.ts
@@ -9,11 +9,12 @@ const METADATA_API = axios.create({
   timeout: TIMEOUT_IN_SECONDS,
 });
 
-export const postValidate = async <MetadataType>(body: MetadataValidationDTO) => {
-  const response = await METADATA_API.post<ValidateMetadataResult<MetadataType>>(
-    `/validate`,
-    body,
-  );
+export const postValidate = async <MetadataType>(
+  body: MetadataValidationDTO,
+) => {
+  const response = await METADATA_API.post<
+    ValidateMetadataResult<MetadataType>
+  >(`/validate`, body);
 
   return response.data;
 };